Comparing commits

Comparing commits

You can use judge to compare benchmark results of two versions of the package.

BenchmarkTools.judge — Function.

judge(pkg::String,
      [target]::Union{String, BenchmarkConfig},
      baseline::Union{String, BenchmarkConfig};
      kwargs...)

Arguments:

pkg - The package to benchmark.

target - What do judge, given as a git id or a BenchmarkConfig. If skipped, use the current state of the package repo.

baseline - The commit / BenchmarkConfig to compare target against.

Keyword arguments:

f - Estimator function to use in the judging.

judgekwargs::Dict{Symbol, Any} - keyword arguments to pass to the judge function in BenchmarkTools

The remaining keyword arguments are passed to benchmarkpkg

Return value:

Returns a BenchmarkJudgement

source

judge(target::BenchmarkResults, baseline::BenchmarkResults, f;
      judgekwargs = Dict())

Judges the two BenchmarkResults in target and baseline using the function f.

Return value

Returns a BenchmarkJudgement

source